moved UserDashboard to pages directory and refactored code to compensate
