+++ frontend/src/routes.js
@@ -3,10 +3,10 @@ import AgileBoard from "./components/regions/AgileBoard";
 import UserActions from "./components/regions/UserActions";
 import GroupCardSummary from "./components/regions/GroupCardSummary";
 import UsersAndGroups from "./components/regions/UsersAndGroups";
-import UserDashboard from "./components/regions/UserDashboard";
 import CardDetails from "./components/regions/CardDetails";
 import Redirector from "./components/regions/Redirector";
 import TeamDashboard from "./components/pages/TeamDashboard";
+import UserDashboard from "./components/pages/UserDashboard";
 
 import { TEAM_PERMISSIONS } from "./constants";
